% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{backend_draws_array}
\alias{backend_draws_array}
\title{Posterior draws of a fit as an iterations x chains x parameters array}
\usage{
backend_draws_array(raw_fit)
}
\arguments{
\item{raw_fit}{a backend-native fit object (an rstan \code{stanfit} or a cmdstanr
\code{CmdStanMCMC}).}
}
\value{
a 3-D array, dimensions iterations x chains x parameters.
}
\description{
Posterior draws of a fit as an iterations x chains x parameters array
}
\examples{
\dontrun{
draws <- backend_draws_array(fit)
dim(draws) # iterations x chains x parameters
}

}
